+++ envpool/sokoban/BUILD
@@ -1,76 +1,75 @@
-# Copyright 2023-2024 FAR AI
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 load("@pip_requirements//:requirements.bzl", "requirement")
 load("@pybind11_bazel//:build_defs.bzl", "pybind_extension")
 
 package(default_visibility = ["//visibility:public"])
 
-# Core C++ libraries
+# 1. Level Loader
+cc_library(
+    name = "level_loader_lib",
+    hdrs = ["level_loader.h", "utils.h"],
+    srcs = ["level_loader.cc"],
+    # if needed, add any includes or linkopts here
+)
+
+# 2. Sokoban Node
+cc_library(
+    name = "sokoban_node_lib",
+    hdrs = ["sokoban_node.h"],        # "level_loader.h" is #included from .h if needed
+    srcs = ["sokoban_node.cc"],
+    deps = [
+        "//third_party/astar_stl:astar_stl_h",
+        ":level_loader_lib",         # only if SokobanNode uses LevelLoader calls
+        # If "utils.h" is used in sokoban_node.cc, you can add a tiny library for it
+        # or just treat it as header-only. If header-only, you may not need an extra dep.
+    ],
+)
+
+# 3. (Optional) If utils.h is purely header-only, you can omit a library. Otherwise:
+# cc_library(
+#     name = "sokoban_utils_lib",
+#     hdrs = ["utils.h"],
+#     # no srcs if purely templates header-only, or "utils.cc" if it exists
+# )
+
+# 4. Sokoban Env
 cc_library(
     name = "sokoban_envpool_lib",
     hdrs = [
-        "level_loader.h",
         "sokoban_envpool.h",
-        "utils.h",
     ],
     srcs = [
         "sokoban_envpool.cc",
     ],
     deps = [
+        # these come from your existing config
         "//envpool/core:async_envpool",
         "//envpool/core:env",
         "//envpool/core:env_spec",
+        ":level_loader_lib",  # needed because we actually instantiate LevelLoader in .cc
+        # If "utils.h" needed a separate library, you'd add it here too
     ],
 )
 
-cc_library(
-    name = "sokoban_node_h",
-    hdrs = [
-        "level_loader.h",
-        "sokoban_node.h",
-        "utils.h",
-    ],
-    deps = ["//third_party/astar_stl:astar_stl_h"],
-)
-
-cc_library(
-    name = "level_loader_lib",
-    hdrs = ["level_loader.h"],
-    srcs = ["level_loader.cc"],
-)
-
-cc_library(
-    name = "sokoban_node_lib",
-    srcs = ["sokoban_node.cc"],
-    deps = [
-        ":sokoban_node_h",
-    ],
-)
-
+# 5. astar_log
 cc_library(
     name = "astar_log_lib",
-    srcs = ["astar_log.cc"],
+    srcs = [
+        "astar_log.cc",  # no main here
+    ],
     deps = [
-        ":level_loader_lib",
         ":sokoban_node_lib",
+        ":level_loader_lib",
+        "//third_party/astar_stl:astar_stl_h",
+        # anything else as needed
     ],
 )
 
-# Binaries
+# The actual CLI tool with the main:
 cc_binary(
     name = "astar_log",
+    srcs = [
+        "astar_log_main.cc",  # <--- main() is here now
+    ],
     deps = [
         ":astar_log_lib",
     ],
@@ -85,35 +84,40 @@ cc_binary(
     ],
 )
 
-# Tests
+# 6. Python tests
 py_test(
     name = "test",
     srcs = ["sokoban_py_envpool_test.py"],
     main = "sokoban_py_envpool_test.py",
     deps = [
         ":registration",
-        ":sokoban",
+        ":sokoban",   # lbry below
         "//envpool",
         requirement("numpy"),
         requirement("pytest"),
     ],
 )
 
+# Now your test doesn't accidentally invoke that main.
 cc_test(
     name = "astar_log_test",
-    srcs = ["astar_log_test.cc"],
+    srcs = [
+        "astar_log_test.cc",
+    ],
     deps = [
         ":astar_log_lib",
         "@com_google_googletest//:gtest_main",
     ],
 )
 
-# Python code
+# 7. Python code and extension
 py_library(
     name = "sokoban",
     srcs = ["__init__.py"],
     data = [":sokoban_envpool.so"],
-    deps = ["//envpool/python:api"],
+    deps = [
+        "//envpool/python:api",
+    ],
 )
 
 py_library(
@@ -124,8 +128,6 @@ py_library(
     ],
 )
 
-
-# Python extension
 pybind_extension(
     name = "sokoban_envpool",
     linkopts = [
@@ -133,6 +135,6 @@ pybind_extension(
     ],
     deps = [
         ":sokoban_envpool_lib",
-        "//envpool/core:py_envpool",
+        "//envpool/core:py_envpool",  # presumably needed for the PyEnvSpec/PyEnvPool macros
     ],
 )
